fingerprint-access-control/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
pydbus==0.6.0
jsonschema==3.2.0
pytest==6.2.4
pytest-cov==2.12.1
Werkzeug==2.0.3